Repair method that builds is-in-list criteria
The method that builds criteria for matching a value in a list was
invoked with "and" so the query was built to match every value in a list
and that is impossible; e.g., access type = 'PR' AND access type = 'PB'.
Also added a test for the scenario so it's not missed again.

Change-Id: I1ab0bd556a75bbeebdf46b5ba045d7e060bbb166
Issue-ID: ACUMOS-1707
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>